src/app/requirements.txt
dash
plotly
pandas
flask
gunicorn
google-cloud-storage
